move Indexer and producer to a separate module
