+++ substrate-query-framework/index-builder/src/index.ts
@@ -1,9 +1,9 @@
 import ISubstrateService, { makeSubstrateService } from './substrate/ISubstrateService';
-import QueryBlockProducer from './QueryBlockProducer';
+import QueryBlockProducer from './indexer/QueryBlockProducer';
 import QueryEventProcessingPack from './QueryEventProcessingPack';
 import QueryEvent from './QueryEvent';
 import QueryEventBlock from './QueryEventBlock';
-import IndexBuilder from './IndexBuilder';
+import IndexBuilder from './indexer/IndexBuilder';
 import QueryNode, { QueryNodeState } from './QueryNode';
 import QueryNodeManager from './QueryNodeManager';
 import { DatabaseManager, SavedEntityEvent, makeDatabaseManager, createDBConnection } from './db';
